+++ src/components/Scheme/SchemeContainer.tsx
@@ -5,7 +5,7 @@ import { enableWalletProvider, getArc } from "arc";
 import classNames from "classnames";
 import Loading from "components/Shared/Loading";
 import withSubscription, { ISubscriptionProps } from "components/Shared/withSubscription";
-import { schemeName, getSchemeIsActive} from "lib/util";
+import { schemeName, getSchemeIsActive} from "lib/schemeUtils";
 import * as React from "react";
 import { BreadcrumbsItem } from "react-breadcrumbs-dynamic";
 import { Helmet } from "react-helmet";
@@ -20,6 +20,7 @@ import { ICrxRewarderProps, getCrxRewarderProps, hasRewarderContract, CrxRewarde
 import ReputationFromToken from "./ReputationFromToken";
 import SchemeInfoPage from "./SchemeInfoPage";
 import SchemeProposalsPage from "./SchemeProposalsPage";
+import SchemeOpenBountyPage from "./SchemeOpenBountyPage";
 import * as css from "./Scheme.scss";
 
 interface IDispatchProps {
@@ -77,7 +78,7 @@ class SchemeContainer extends React.Component<IProps, IState> {
 
   private schemeInfoPageHtml = (props: any) => <SchemeInfoPage {...props} daoState={this.props.daoState} scheme={this.props.data[0]} />;
   private schemeProposalsPageHtml = (isActive: boolean) => (props: any) => <SchemeProposalsPage {...props} isActive={isActive} daoState={this.props.daoState} currentAccountAddress={this.props.currentAccountAddress} scheme={this.props.data[0]} />;
-  private contributionsRewardExtTabHtml = () => (props: any) => 
+  private contributionsRewardExtTabHtml = () => (props: any) =>
   {
     if (!this.state.crxListComponent) {
       return null;
@@ -115,21 +116,27 @@ class SchemeContainer extends React.Component<IProps, IState> {
 
     const proposalsTabClass = classNames({
       [css.proposals]: true,
-      [css.active]: !this.props.location.pathname.includes("info") && !this.props.location.pathname.includes("crx"),
+      [css.active]: !this.props.location.pathname.includes("info") && !this.props.location.pathname.includes("crx") && !this.props.location.pathname.includes("open"),
     });
     const infoTabClass = classNames({
       [css.info]: true,
       [css.active]: this.props.location.pathname.includes("info"),
     });
+    const openBountiesTabClass = classNames({
+      [css.openbounty]: true,
+      [css.active]: this.props.location.pathname.includes("openbounties"),
+    });
+
     const crxTabClass = classNames({
       [css.crx]: true,
       [css.active]: this.props.location.pathname.includes("crx"),
     });
     const schemeFriendlyName = schemeName(schemeState, schemeState.address);
-    
+
+
     return (
       <div className={css.schemeContainer}>
-     
+
         <BreadcrumbsItem to={`/dao/${daoAvatarAddress}/scheme/${schemeId}`}>{schemeFriendlyName}</BreadcrumbsItem>
         <Helmet>
           <meta name="description" content={daoState.name + " | " + schemeState.name + " proposals | Managed on Alchemy by DAOstack"} />
@@ -144,6 +151,11 @@ class SchemeContainer extends React.Component<IProps, IState> {
 
           <div className={css.schemeMenu}>
             <Link className={proposalsTabClass} to={`/dao/${daoAvatarAddress}/scheme/${schemeId}/proposals/`}>Proposals</Link>
+
+            { // if Bounties Scheme, create new tab
+              (schemeName(schemeState, schemeState.address) === "Standard Bounties") &&
+              <Link className={openBountiesTabClass} to={`/dao/${daoAvatarAddress}/scheme/${schemeId}/openbounties/`}>Open Bounties</Link>
+            }
             <TrainingTooltip placement="top" overlay={"Learn about the protocol parameters for this scheme"}>
               <Link className={infoTabClass} to={`/dao/${daoAvatarAddress}/scheme/${schemeId}/info/`}>Information</Link>
             </TrainingTooltip>
@@ -172,6 +184,9 @@ class SchemeContainer extends React.Component<IProps, IState> {
         </Sticky>
 
         <Switch>
+
+          <Route exact path="/dao/:daoAvatarAddress/scheme/:schemeId/openbounties"
+            render={(props) => <SchemeOpenBountyPage {...props} daoAvatarAddress={daoAvatarAddress} scheme={schemeState} />} />
           <Route exact path="/dao/:daoAvatarAddress/scheme/:schemeId/info" render={this.schemeInfoPageHtml} />
           {
             this.state.crxRewarderProps ?