+++ src/components/Shared/PreTransactionModal.tsx
@@ -102,7 +102,6 @@ class PreTransactionModal extends React.Component<IProps, IState> {
     // Do action on Enter key press
     if (e.keyCode === 13) {
       if (actionType === ActionTypes.StakeFail || actionType === ActionTypes.StakePass) {
-        console.log(this.state.stakeAmount, fromWei(currentAccountGens), this.state.stakeAmount > 0);
         if (this.state.stakeAmount > 0 && this.state.stakeAmount <= fromWei(currentAccountGens)) {
           this.handleClickAction();
         }
@@ -303,7 +302,7 @@ class PreTransactionModal extends React.Component<IProps, IState> {
               <div className={css.transactionIcon}>{icon}</div>
               <div className={css.transactionInfo}>
                 <span className={css.transactionType}>{transactionType}</span>
-                { !multiLineMsg ? <span className={css.titleSeparator}>|</span>: "" }
+                { (!multiLineMsg && secondaryHeader) ? <span className={css.titleSeparator}>|</span>: "" }
                 <span className={css.secondaryHeader}>{secondaryHeader}</span>
                 <div className={css.transactionEffect}>
                   {effectText}